+++ micro_academico/src/main/java/com/example/micro_academico/ObjetivoDeAprendizaje/repository/ObjetivoAprendizajeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ObjetivoAprendizajeRepository extends JpaRepository<ObjetivoAprendizaje, Long> {
+public interface ObjetivoAprendizajeRepository extends JpaRepository<ObjetivoAprendizaje, Integer> {
 }
